# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM API Documentation",
        "description": "CRM API Documentation",
        "contact": {
            "email": "ct.ranatunga@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "CRM API Server"
        }
    ],
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "AdminLogin"
                ],
                "summary": "AdminLogin",
                "security": [],
                "description": "Returns access token",
                "operationId": "adminLogin",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/crm/customers/{filters}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get list of customers",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Returns list of customers",
                "operationId": "getCustomerList",
                "parameters": [
                    {
                        "name": "filters",
                        "in": "path",
                        "description": "Search keyword or phone numbers",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "se=chamara"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/crm/{customerId}/customer": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get Customer by uuid",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Returns customer by uuid",
                "operationId": "getCustomer",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "path",
                        "description": "Customer uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/crm/customer": {
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Create Customer",
                "description": "Create Customer",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "operationId": "createCustomer",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Customer"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/crm/{customerUuid}/customer": {
            "put": {
                "tags": [
                    "Customers"
                ],
                "summary": "Update Customer",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Update Customer",
                "operationId": "updateCustomer",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Customer"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "customerUuid",
                        "in": "path",
                        "description": "Customer uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/crm/{customerDeleteUuid}/customer": {
            "delete": {
                "tags": [
                    "Customers"
                ],
                "summary": "Delete Customer by uuid",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Delete customer by uuid",
                "operationId": "deleteCustomer",
                "parameters": [
                    {
                        "name": "customerDeleteUuid",
                        "in": "path",
                        "description": "Customer uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/crm/customers/import": {
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Import customers by csv file",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Import customers by csv file",
                "operationId": "importCustomers",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Import"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Customer": {
                "type": "object",
                "required": [
                    "first_name",
                    "last_name",
                    "email"
                ],
                "properties": {
                    "first_name": {
                        "type": "string",
                        "example": "doggie"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "doggie1"
                    },
                    "email": {
                        "type": "string",
                        "example": "doggie@web.com"
                    },
                    "phone_numbers": {
                        "type": "array",
                        "xml": {
                            "name": "phone_numbers"
                        },
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "User": {
                "type": "object",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "admin@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password"
                    }
                }
            },
            "Import": {
                "type": "object",
                "properties": {
                    "file": {
                        "type": "string",
                        "format": "binary"
                    }
                   
                },
                "required": ["file"]
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Passport security.",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "CRM",
            "description": "API Endpoints of CRM"
        }
    ],
    "security": [
        []
    ]
}